refactor: move `TimeStampModel` from models/custom_user.py to models/abstract.py
