+++ blog_app/models/custom_user.py
@@ -1,15 +1,7 @@
 from django.contrib.auth.models import User, Group, Permission
 from django.db import models
 from django.contrib.auth.models import AbstractUser
-
-
-class TimeStampModel(models.Model):
-    created_at = models.DateTimeField(auto_now_add=True)
-    updated_at = models.DateTimeField(auto_now=True)
-    deleted_at = models.DateTimeField(null=True, blank=True)
-
-    class Meta:
-        abstract = True
+from blog_app.models.abstract import TimeStampModel
 
 
 class CustomUser(TimeStampModel, AbstractUser):
